Move api-calls
