+++ frontend/src/components/pages/Auth.js
@@ -3,7 +3,7 @@ import AuthContext from '../context/auth-context';
 import StyledFirebaseAuth from 'react-firebaseui/StyledFirebaseAuth';
 import firebase from 'firebase/app';
 import 'firebase/auth';
-import { fetchNoAuth } from '../api-calls/Fetch';
+import { fetchNoAuth } from '../../api-calls/Fetch';
 
 import "./Auth.css";
 import 'firebaseui/dist/firebaseui.css';
