+++ frontend/src/components/pages/Home.js
@@ -6,7 +6,7 @@ import { Redirect } from 'react-router-dom';
 import Modal from '../Modal/Modal';
 import Backdrop from '../Backdrop/Backdrop';
 import AuthContext from '../context/auth-context';
-import { fetchRun } from '../api-calls/Fetch';
+import { fetchRun } from '../../api-calls/Fetch';
 import './Home.css'
 
 /**
